chore(tests): centralize WP test bootstrap
